+++ src/interfaces/RouteProps.interface.tsx
@@ -1,4 +1,4 @@
-import type { RouteParameters } from '../types/RouteProps.type.tsx';
+import type { RouteParameters } from '../types/RouteParameters.type.tsx';
 
 export interface RouteProps {
     path: string;
